Refactor: move blackboard to its own module for cleanliness.
