+++ src/rust/uisp_integration/src/main.rs
@@ -7,14 +7,14 @@ mod errors;
 pub mod ip_ranges;
 mod strategies;
 pub mod uisp_types;
+mod blackboard;
 
 use crate::errors::UispIntegrationError;
 use crate::ip_ranges::IpRanges;
-use lqos_bus::BlackboardSystem;
 use lqos_config::Config;
-use serde::Serialize;
 use tokio::time::Instant;
 use tracing::{error, info};
+pub use blackboard::*;
 
 /// Start the tracing/logging system
 fn init_tracing() {
@@ -35,43 +35,6 @@ fn check_enabled_status(config: &Config) -> Result<(), UispIntegrationError> {
     }
 }
 
-pub async fn blackboard(subsystem: BlackboardSystem, key: &str, value: &str) {
-    let Ok(config) = lqos_config::load_config() else {
-        return;
-    };
-    if !config.long_term_stats.use_insight.unwrap_or(false) {
-        return;
-    }
-    let req = vec![lqos_bus::BusRequest::BlackboardData {
-        subsystem,
-        key: key.to_string(),
-        value: value.to_string(),
-    }];
-    let _ = lqos_bus::bus_request(req).await;
-}
-
-pub async fn blackboard_blob<T: Serialize>(key: &str, value: T) -> anyhow::Result<()> {
-    let config = lqos_config::load_config()?;
-    if !config.long_term_stats.use_insight.unwrap_or(false) {
-        return Ok(());
-    }
-    let blob = serde_cbor::to_vec(&value)?;
-    let chunks = blob.chunks(1024 * 16);
-    info!(
-        "Blob {key} is {} bytes long, split into {} chunks",
-        blob.len(),
-        chunks.len()
-    );
-    for (i, chunk) in chunks.enumerate() {
-        let req = vec![lqos_bus::BusRequest::BlackboardBlob {
-            tag: key.to_string(),
-            part: i,
-            blob: chunk.to_vec(),
-        }];
-        let _ = lqos_bus::bus_request(req).await;
-    }
-    Ok(())
-}
 
 #[tokio::main]
 async fn main() -> Result<(), UispIntegrationError> {
